requirements.txt
flask>=2.3,<4
flask-cors>=4.0,<5
psycopg2-binary>=2.9,<3
authlib>=1.2,<2
python-dotenv>=1.0,<2
pymongo>=4.6,<5
werkzeug>=2.3,<4
